Edit
Data yang muncul di dataadmincafe sekarang tidak ada lagi user sekarang yg sedang login
Data admin yang sedang login di admincafe dapat di edit melalui admincafe sekarang

+++ dataadmincafe.php
@@ -1,16 +1,19 @@
 <?php
 	$con = mysqli_connect("localhost", "root", "", "dbcafe");
+	session_start();
 	$result = mysqli_query($con, "SELECT * FROM admincafe");
 	while ($row = mysqli_fetch_array($result))
 	{
-		echo "Nama: ".$row["nama"]."<br>";
-		echo "Tanggal Lahir: ".$row["tgllahir"]."<br>";
-		echo "Jenis Kelamin: ".$row["jk"]."<br>";
-		echo "Alamat: ".$row["alamat"]."<br>";
-		echo '<a href="editadmincafe.php? id='.$row[0].'&name='.$row[1].'&tgllahir='.$row[2].'&jk='.$row[3].'&alamat='.$row[4].'"><button> Edit </button></a>';
-		echo '<a href="deleteadmincafe.php? id='.$row[0].'""><button>Hapus</button></a>';
+		if ($row['id'] != $_SESSION['idnow']){
+			echo "Nama: ".$row["nama"]."<br>";
+			echo "Tanggal Lahir: ".$row["tgllahir"]."<br>";
+			echo "Jenis Kelamin: ".$row["jk"]."<br>";
+			echo "Alamat: ".$row["alamat"]."<br>";
+			echo '<a href="editadmincafe.php? id='.$row[0].'&name='.$row[1].'&tgllahir='.$row[2].'&jk='.$row[3].'&alamat='.$row[4].'&username='.$row[5].'"><button> Edit </button></a>';
+			echo '<a href="deleteadmincafe.php? id='.$row[0].'""><button>Hapus</button></a>';
 
-		echo "<br />";
+			echo "<br />";
+		}
 	}
 ?>
